requirements.txt
PyPDF2
python-docx
moviepy
pdf2pptx
